requirements.txt
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.13
langchain-chroma>=0.0.10
chromadb>=0.4.22
sentence-transformers>=2.2.2
transformers>=4.36.2
torch>=2.1.2
python-dotenv>=1.0.0
pypdf>=3.17.1
pydantic>=2.5.2
fastapi>=0.104.1
uvicorn>=0.24.0
streamlit>=1.29.0